% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_gof.R
\name{chisq_gof}
\alias{chisq_gof}
\title{chisq_gof}
\usage{
chisq_gof(vector_1, vector_2, reps)
}
\arguments{
\item{vector_1}{\itemize{
\item Numeric vector
}}

\item{vector_2}{\itemize{
\item Numeric vector
}}

\item{reps}{\itemize{
\item Number of simulations.
}}
}
\value{

}
\description{
chisq_gof
}
\examples{
x <- c(rep(1, 15), rep(2, 5))
y <- c(1, 2)

chisq_gof(x, y, reps = 10000)

}
