% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_gof.R
\name{chisq_gof}
\alias{chisq_gof}
\title{Simulated p-values for Chi Squared goodness-of-fit}
\usage{
chisq_gof(x, y, reps = 10000)
}
\arguments{
\item{x}{\itemize{
\item a numeric vector.
}}

\item{y}{\itemize{
\item a numeric vector representing the reference distribution.
}}

\item{reps}{\itemize{
\item an integer specifying the number of simulations. The default is set to 10,000
which is appropriate for exploratory analysis. A higher number of simulations
should be selected for final analysis.

}}
}
\value{

}
\description{

Simulated p-values for Chi Squared goodness-of-fit

}
\examples{
x <- c(rep(1, 15), rep(2, 5))
y <- c(1, 2)

chisq_gof(x, y, reps = 10000)

}
