% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_gof.R
\name{g_gof}
\alias{g_gof}
\title{Simulated p-values for G^2 ("log-likelihood-ratio") goodness-of-fit}
\usage{
g_gof(x, p, reps = 10000)
}
\arguments{
\item{x}{\itemize{
\item a numeric vector that contains counts for each bin/category.
}}

\item{p}{\itemize{
\item a numeric vector that contains probabilities for each bin/category. .
}}

\item{reps}{\itemize{
\item an integer specifying the number of Monte Carlo simulations to perform.
The default is set to 10,000 which is appropriate for exploratory analysis.
A higher number of simulations should be selected for final analysis.
}}
}
\value{
A number representing the p-value.
}
\description{
Simulated p-values for G^2 ("log-likelihood-ratio") goodness-of-fit
}
\examples{
x <- c(15, 36, 17)
p <- c(0.25, 0.5, 0.25)

g_gof(x, p)

}
